refactor(ui): modularize chat loop pipeline
